+++ packages/cli/src/commands/migrate.test.ts
@@ -3,9 +3,8 @@
 
 import fs from 'fs';
 import path from 'path';
-import {promisify} from 'util';
 import {makeTempDir} from '@subql/common';
-import rimraf from 'rimraf';
+import {rimraf} from 'rimraf';
 import {DEFAULT_SUBQL_MANIFEST} from '../constants';
 import Migrate from './migrate';
 
@@ -20,7 +19,7 @@ describe('Intergration test - Migrate', () => {
   afterAll(async () => {
     try {
       if (!projectDir) return;
-      await promisify(rimraf)(projectDir);
+      await rimraf(projectDir);
     } catch (e) {
       console.warn('Failed to clean up tmp dir after test', e);
     }