+++ packages/cli/src/controller/add-chain-controller.spec.ts
@@ -4,11 +4,10 @@
 import fs, {existsSync} from 'fs';
 import os from 'os';
 import path from 'path';
-import {promisify} from 'util';
 import {DEFAULT_MULTICHAIN_MANIFEST} from '@subql/common';
 import {MultichainProjectManifest, ProjectManifestV1_0_0} from '@subql/types-core';
 import * as yaml from 'js-yaml';
-import rimraf from 'rimraf';
+import {rimraf} from 'rimraf';
 import {YAMLSeq} from 'yaml';
 import {loadMultichainManifest, validateAndAddChainManifest} from './add-chain-controller';
 
@@ -135,7 +134,7 @@ describe('MultiChain - ADD', () => {
 
   afterEach(async () => {
     try {
-      await promisify(rimraf)(projectDir);
+      await rimraf(projectDir);
     } catch (e) {
       console.warn('Failed to clean up tmp dir after test', e);
     }