+++ packages/cli/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update inquirer dependencies (#2501)
 - Improve endpoint type to be based on core type (#2511)
+- Update dependencies (#2518)
 
 ### Fixed
 - Not being able to select a custom repo when initializing a project (#2501)